+++ ve-silo/contracts/gauges/interfaces/ISiloLiquidityGauge.sol
@@ -6,7 +6,7 @@ interface ISiloLiquidityGauge {
     // solhint-disable func-name-mixedcase
     // solhint-disable func-param-name-mixedcase
     // solhint-disable var-name-mixedcase
-    function balance_updated_for_users(
+    function afterTokenTransfer(
         address _user1,
         uint256 _user1_new_balancer,
         address _user2,
@@ -18,8 +18,12 @@ interface ISiloLiquidityGauge {
 
     function user_checkpoint(address _addr) external returns (bool);
 
-    /// @notice Returns ERC-20 Balancer handler
-    function bal_handler() external view returns (address);
+    /// @notice Returns a silo hook receiver
+    function hook_receiver() external view returns (address);
+    /// @notice Returns a silo share token
+    function share_token() external view returns (address);
+    /// @notice Returns a silo
+    function silo() external view returns (address);
     /// @notice Get the timestamp of the last checkpoint
     function integrate_checkpoint() external view returns (uint256);
     /// @notice ∫(balance * rate(t) / totalSupply(t) dt) from 0 till checkpoint